# Conflicts:
#	Projects/Server/Mobiles/Mobile.cs
#	Projects/UOContent/Engines/Treasures of Tokuno/TreasuresOfTokuno.cs
#	Projects/UOContent/Items/Talismans/BaseTalisman.cs
#	Projects/UOContent/Items/Weapons/Abilities/MortalStrike.cs
#	Projects/UOContent/Mobiles/PlayerMobile.cs
#	Projects/UOContent/Spells/Base/MagerySpell.cs
#	Projects/UOContent/Spells/Bushido/SamuraiSpell.cs
#	Projects/UOContent/Spells/Chivalry/NobleSacrifice.cs
#	Projects/UOContent/Spells/Mysticism/SpellPlagueSpell.cs
#	Projects/UOContent/Spells/Necromancy/BloodOathSpell.cs
#	Projects/UOContent/Spells/Necromancy/CorpseSkin.cs
#	Projects/UOContent/Spells/Necromancy/CurseWeapon.cs
#	Projects/UOContent/Spells/Necromancy/MindRot.cs
#	Projects/UOContent/Spells/Necromancy/NecromancerSpell.cs
#	Projects/UOContent/Spells/Necromancy/Wither.cs
#	Projects/UOContent/Spells/Ninjitsu/AnimalForm.cs
#	Projects/UOContent/Spells/Ninjitsu/DeathStrike.cs
#	Projects/UOContent/Spells/Ninjitsu/KiAttack.cs
#	Projects/UOContent/Spells/Ninjitsu/MirrorImage.cs
#	Projects/UOContent/Spells/Ninjitsu/NinjaSpell.cs
#	Projects/UOContent/Spells/Ninjitsu/SurpriseAttack.cs
#	Projects/UOContent/Spells/Spellweaving/ArcaneCircle.cs
#	Projects/UOContent/Spells/Spellweaving/GiftOfRenewal.cs

+++ Projects/UOContent/Multis/Houses/PreviewHouse.cs
@@ -21,7 +21,7 @@ namespace Server.Multis
 
                 if (entry.Flags == 0)
                 {
-                    Item item = new Static((int)entry.ItemId);
+                    Item item = new Static(entry.ItemId);
 
                     item.MoveToWorld(new Point3D(X + entry.OffsetX, Y + entry.OffsetY, Z + entry.OffsetZ), Map);
